requirements.txt
asgiref==3.7.2
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
Django==4.2.2
django-ckeditor==6.5.1
django-cloudinary-storage==0.3.0
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
psycopg2-binary==2.9.6
python-decouple==3.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
webencodings==0.5.1
whitenoise==6.5.0
